pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["Hossam Hassan <61355308+hossam4455@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
Pillow = "^9.2.0"
django-imagekit = "^4.1.0"
django-model-utils = "^4.2.0"
django-widget-tweaks = "^1.4.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
